requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
torch==2.4.1
Pillow==10.4.0